Removed class "X4T0U" from being selected in results
If this class is selected, the results will add the inline 'Video' section which can appear in the main Google results page sometimes. We do not want to add this yet as we are only interested in the organic results, but we may add it later as a future improvement to the project.